Added package for questions and set up the start page of the app
"questions" is a package containing classes for working with questions to
be used by the app. Question is a base class off which text and image
question classes are based, each having different methods for handling the
type of the question. QuestionFactory just builds an array of Questions
(which are generated statically for now) and returns it to the user of the
factory.

The start page activity contains a button to start the quiz: the button
does nothing as of now but will lead to the first question of the quiz as
more of the app is implemented.