Implemented the entire functionality of the app.
* The QuestionFactory interface is modified to have two iterators - one for image based questions and another for text based questions.
* TextQuestion and ImageQuestion are modified to properly integrate with
rest of the application: TextQuestion stores a String[] of options instead
of an ArrayList as this is easier to work with, and ImageQuestion stores
the resource ID of the image to be used.
* ProgressFragment exposes an interface implemented by QuizPlayActivity,
and is used to notify the latter that the question is to be progressed.
* QuizPlayActivity loads the question when the activity gains focus or
when the progress fragment button is clicked. If all questions are
exhuasted, it launches QuizEndActivity which displays the score and gives
options to end the app or restart the game.
* QuizPlayActivity is removed from the back stack once it ends to prevent
someone from going back to change their answers. Instead, pressing back
from QuizEndActivity loads the start page.